+++ seq2science/util.py
@@ -201,6 +201,7 @@ def sieve_bam(configdict):
         configdict.get("min_mapping_quality", 0) > 0
         or configdict.get("tn5_shift", False)
         or configdict.get("remove_blacklist", False)
+        or configdict.get("filter_on_size", False)
         or configdict.get("remove_mito", False)
     )