+++ tests/dag_tests.sh
@@ -1,4 +1,5 @@
 # to run these tests locally:
+#   mamba install toml --yes
 #   python setup.py develop
 #   bash ./tests/dag_tests.sh TEST
 
@@ -210,6 +211,16 @@ if [ $1 = "atac-seq" ]; then
   printf "\nmultiqc report\n"
   seq2science run atac-seq -n --configfile tests/alignment/default_config.yaml --snakemakeOptions quiet=True config={create_qc_report:True,trimmer:trimgalore} | tee tests/local_test_results/${1}_dag
   assert_rulecount $1 fastqc 4
+
+  printf "\natac-seq template length filter\n"
+  seq2science run atac-seq -n --configfile tests/alignment/default_config.yaml --snakemakeOptions quiet=True config={create_qc_report:True} | tee tests/local_test_results/${1}_dag
+  assert_rulecount $1 mark_duplicates 1
+
+  seq2science run atac-seq -n --configfile tests/alignment/default_config.yaml --snakemakeOptions quiet=True config={create_qc_report:True,min_template_length:150} | tee tests/local_test_results/${1}_dag
+  assert_rulecount $1 mark_duplicates 2
+
+  seq2science run atac-seq -n --configfile tests/alignment/default_config.yaml --snakemakeOptions quiet=True config={create_qc_report:True,max_template_length:150} | tee tests/local_test_results/${1}_dag
+  assert_rulecount $1 mark_duplicates 2
 
   printf "\ncustom assembly\n"
   seq2science run atac-seq -n --configfile tests/$WF/macs2.yaml --snakemakeOptions quiet=True config={custom_genome_extension:tests/local_test_results/ERCC92.fa} | tee tests/local_test_results/${1}_dag